use Link instead of a tag
'use client' is needed only if you’re using React Server Components or building a library compatible with them. <Link> is a built-in component that extends the HTML <a> tag to provide prefetching and client-side navigation between routes. It is the primary and recommended way to navigate between routes in Next.js